Update the "test-scripts" dir & fix lint errors
