+++ test-scripts/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "protocol-cosmwasm-launcher",
+  "name": "protocol-cosmwasm-test-scripts",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
   "scripts": {
+    "lint": "eslint . --ext .ts",
     "test": "ts-node ./src/index.ts",
     "test:setup-mixer": "npm run test --mode=testnet_setup_mixer",
     "test:test-mixer": "npm run test --mode=testnet_test_mixer",
@@ -16,5 +17,10 @@
     "fs": "^0.0.1-security",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
+    "@typescript-eslint/parser": "^5.19.0",
+    "eslint": "^8.13.0"
   }
 }
