+++ test-scripts/src/config.ts
@@ -1,0 +1,21 @@
+
+export default {
+    // Testnet bombay-12
+    networkInfo: {
+        lcd_url: 'https://bombay.stakesystems.io', // backup: 'https://bombay-lcd.terra.dev'
+        chainId: 'bombay-12',
+        gasPrice: 0.15,
+        gasAdjustment: 1.4,
+    },
+
+    mnemonicKeys: {
+        deployer: "",
+    },
+
+    contracts: {
+        cw20: "terra1340t6lqq6jxhm8d6gtz0hzz5jzcszvm27urkn2",
+        mixer: "",
+        anchor: "",
+        vanchor: "",
+    },
+};
