+++ test-scripts/wasm_contracts/cosmwasm_anchor.wasm
