+++ test-scripts/wasm_contracts/cosmwasm_mixer.wasm
